% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{all_pra_prices}
\alias{all_pra_prices}
\title{Collect All PRA prices in All years}
\usage{
all_pra_prices(yr1950 = 1950, readdisk = TRUE)
}
\description{
we use this function to get prices for all available PRAs
in the country, i.e. not matching to our 200/100 city samples.
we use those prices for the neighborhood IV exercise and for plotting
purposes.
}
