% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities-defs.R
\name{bboxes_top200}
\alias{bboxes_top200}
\title{top 200 city bounding boxes}
\usage{
bboxes_top200()
}
\description{
manually defines bounding boxes for areas containing the
top 100 french cities in WGS84 lat-lon (google maps) CRS.
Also adds the centre for Paris (notre dame) manually in WGS84 lat-lon.
}
