% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet.R
\name{census_add_toutain}
\alias{census_add_toutain}
\title{Complement Census Data with Toutain}
\usage{
census_add_toutain()
}
\description{
This function complements INSEE population count data with Toutain's series.
The output of this is our measure for \emph{Population} in the model.
}
\details{
Source: Toutain (1991) Tableau 1: Evolution de la Population des menages agricoles de 1789 a 1968
}
