% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-areas.R
\name{city_ranking}
\alias{city_ranking}
\title{get a ranking of cities by size}
\usage{
city_ranking(p, yr = 2015, topn = 30, overwrite = FALSE)
}
\description{
this is somewhat arbitrary as it includes communes which are clearly part of larger agglomerations. the population column is not the final product, in other words.
excludes boulogne and versaille, villeurbane and La Reunion
}
