% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLC.R
\name{CLC_bboxes}
\alias{CLC_bboxes}
\title{Crop CLC 2018 raster to 2015 French Cities}
\usage{
CLC_bboxes(overwrite = FALSE)
}
\description{
load CLC raster and cut out our bounding boxes of French cities.
Boxes are obtained via \code{\link{measure_cities}}
}
