% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLC.R
\name{CLC_measure}
\alias{CLC_measure}
\title{Measure Landuse Around Cities with CLC}
\usage{
CLC_measure(overwrite = TRUE)
}
\description{
takes cropped CLC rasters from \code{\link{CLC_bboxes}} for top 100 cities and counts the proportion of land
outside city falling into each category. We plot the top 15 categories of the CLC classification and provide our own aggregation of labels into 11 exhaustive classes of land use.
}
