% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{combine_measures}
\alias{combine_measures}
\title{Combine City Measures}
\usage{
combine_measures(overwrite = FALSE, cutoff = 30)
}
\arguments{
\item{overwrite}{TRUE/FALSE}

\item{cutoff}{numeric value above which classify as urban}
}
\description{
combine satellite measures from \code{\link{measure_cities}} and manual measures
from \code{\link{pop_1950_2}} to create the final output
dataset.
}
