% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris-commute.R
\name{commute_get_active}
\alias{commute_get_active}
\title{Compute Active Population for Paris AU}
\usage{
commute_get_active(noagri = TRUE)
}
\description{
Compute Active Population for Paris AU
}
\details{
Take our own delineation of the Parisian Urban area
from \code{\link{get_paris_AU}} and merge with INSEE population data for each commune.
For 1975 and 1990 we use our own delinations, for 2000 and 2010 we have offical classifications.
}
