% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris-commute.R
\name{commute_paris_AU}
\alias{commute_paris_AU}
\title{Get Paris Aire Urbaine 1975, 1990, 2000}
\usage{
commute_paris_AU(overwrite = FALSE)
}
\description{
uses our method from \code{\link{measure_cities}} to determine
which INSEE communes are part of the parisian urban area. This
is an approximation only based on contiguous urban area (it does not
consider actual commuting patterns). The first official Urban Area classification
is available in 2010.
}
\details{
We convert our raster data into an sf vector graphics object,
and intersect it with the communes shapefile of INSEE.
}
