% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{crop_smod}
\alias{crop_smod}
\title{Crop GHS Settlement Model for One City}
\usage{
crop_smod(bbox, r, yr, cutoff = 20, overwrite = FALSE)
}
\arguments{
\item{yr}{for which year to do that}
}
\description{
Crop GHS Settlement Model for One City
}
