% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{cutoff_sensitivity}
\alias{cutoff_sensitivity}
\title{Cutoff Sensitivity Analysis for \code{measure_cities}}
\usage{
cutoff_sensitivity(cuts = c(25, 30, 35, 50), overwrite = FALSE)
}
\arguments{
\item{cuts}{desired cutoff levels

compute output measures for different values of \code{cutoff}}
}
\description{
Cutoff Sensitivity Analysis for \code{measure_cities}
}
