% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dboxdir}
\alias{dboxdir}
\title{Get Dropbox Root}
\usage{
dboxdir()
}
\description{
This returns the root of our shared dropbox folder on your machine.
On my computer this would be \verb{~/Dropbox/research/LandUse}
}
