% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{dist_from_center}
\alias{dist_from_center}
\title{Compute Distance of Fringe from City Center}
\usage{
dist_from_center(overwrite = FALSE, ndists = 51)
}
\arguments{
\item{overwrite}{TRUE/FALSE whether to load from disk or recreate}

\item{ndists}{numeric value indicating how many distance bins}
}
\value{
returns a data.table with distance, density by city and year.
}
\description{
Read cut-out city rasters from \code{\link{measure_cities}} and compute
distance of each point in raster to the point called \emph{city center}.
City Center is classified by IGN as \emph{Chef Lieu} of a place.
}
