% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{exp_decay}
\alias{exp_decay}
\title{Exponential Decay Parameters for All Cities}
\usage{
exp_decay(by_year = FALSE, cutoff = NULL)
}
\arguments{
\item{by_year}{by default FALSE}
}
\description{
uses function \code{\link{exp_decay_single}} to compute an exponential decay model
for each city in the sample.
}
