% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{exp_decay_linear}
\alias{exp_decay_linear}
\title{Exponential Decay Model}
\usage{
exp_decay_linear(yr = 2015)
}
\description{
estimates the model log(D_{lk}) = log(a_k) + b_k lk where lk is a location in
city k.
}
