% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{exp_decay_single}
\alias{exp_decay_single}
\title{Estimate Exponential Density Decay Model}
\usage{
exp_decay_single(d, by_year = TRUE)
}
\description{
Takes data from \code{\link{dist_from_center}} and computes an exponential decay
model for a single city.
}
