% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{exp_decay_summary}
\alias{exp_decay_summary}
\title{Summary For Exponential Decay}
\usage{
exp_decay_summary(by_year = TRUE)
}
\description{
takes output of \code{\link{exp_decay}} and summarises the
distribution of estimated exponential decay coefficients. 27 out of 100
cities do not converge for the estimation and are dropped from the sample.
}
