% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{export_pop_area}
\alias{export_pop_area}
\title{Main Output File}
\usage{
export_pop_area(topn = 100)
}
\description{
reads output of manual and satellite measures via \code{\link{read_output}}
and amends Paris Population to also include the Seine department as well before WW2.
}
\details{
The csv output of this file is used in the model.
}
