% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure5}
\alias{figure5}
\title{Produce Figure 5}
\usage{
figure5()
}
\description{
Uses 100 city sample and makes 2 scatters for rural price vs urban density
and urban population vs urban density, both for year 2000.
}
