% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{french_cities}
\alias{french_cities}
\title{Cut Out rasters for French cities}
\usage{
french_cities()
}
\arguments{
\item{yr}{for which year to do that
France UCDB: Cities

List of French cities in study}
}
\description{
Crop GHS Settlement Model for One City
}
