% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet.R
\name{get_manuals}
\alias{get_manuals}
\title{Get Manual Measurements}
\usage{
get_manuals(reload = FALSE)
}
\arguments{
\item{reload}{TRUE/FALSE whether to fetch from google drive or from disk

Notice in google drive sheet area2016 = 0 or \code{is.na(which_city_2016)} means the city historic has been absorbed
by another city in modern times. Hence we do not consider such a city.}
}
\description{
Get Manual Measurements
}
