% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{get_PRA_result_name_1999}
\alias{get_PRA_result_name_1999}
\title{Link PRA names to CODGEO in 1999+}
\usage{
get_PRA_result_name_1999(result, pra)
}
\description{
This function attemps to allocate cities to relevant PRA IDs.
}
\details{
Should probably rely on \code{\link[=PRA_xw]{PRA_xw()}} instead here.

Auvergne Alpes Rhones map: https://draaf.auvergne-rhone-alpes.agriculture.gouv.fr/IMG/pdf/no9_Analyses_VVT_2015_enligne_cle87d384.pdf
}
