% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{get_seine_marne_1990}
\alias{get_seine_marne_1990}
\title{For Seine et Marne area 1990}
\usage{
get_seine_marne_1990()
}
\description{
For the Seine et Marne PRAs disaggregated sales data are missing in 1990
and we only observe rental values (\emph{valeur louee}). We impute sales prices
by using the price to rent ratio at the departmental level. We get the
rental values from \code{\link[=read_terre_labourable_louee_1990]{read_terre_labourable_louee_1990()}}.
}
\seealso{
\code{\link[=get_seine_marne_2000]{get_seine_marne_2000()}}
}
