% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{get_seine_marne_2000}
\alias{get_seine_marne_2000}
\title{Seine et Marne PRA Prices from Departmental Price 2000}
\usage{
get_seine_marne_2000()
}
\description{
For the Seine et Marne PRAs disaggregated sales data are missing in 2000
and we only observe rental values (\emph{valeur louee}). We impute sales prices
by using the price to rent ratio at the departmental level. We get the
rental values from \code{\link[=read_terre_labourable_louee_1999_2000]{read_terre_labourable_louee_1999_2000()}}.
}
