% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{merge_agreste_pra_ref}
\alias{merge_agreste_pra_ref}
\title{Merge PRA Referentiel with Agreste PRA Prices}
\usage{
merge_agreste_pra_ref()
}
\description{
Want to associate a unique PRA_code to each row in Agreste Prices
}
\details{
Attempt to automatically perform match. function deprecated, as we use the manual
matching provided in \code{PRA_xw}.
}
