% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities-defs.R
\name{merge_centers}
\alias{merge_centers}
\title{add center of cities to bounding boxes}
\usage{
merge_centers(overwrite = FALSE)
}
\description{
all except paris (which has center of each arrondissement instead)
}
