% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{paris_central_density}
\alias{paris_central_density}
\title{Generate Central Paris Density}
\usage{
paris_central_density()
}
\description{
We know both area and population for arrondissements 1-6 in Paris
}
\details{
We call those \emph{central paris}. We will use this data to confront with \emph{average Paris}.
}
