% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{paris_pra_1975_labour}
\alias{paris_pra_1975_labour}
\title{Paris land value 1975}
\usage{
paris_pra_1975_labour(tl0, keep_pras = FALSE)
}
\description{
terre labourable only
}
\details{
We take all PRA directly adjacent to paris as well as CEINTURE DE PARIS as
on this map: https://driaaf.ile-de-france.agriculture.gouv.fr/IMG/pdf/RA_R11_cle4ce9b1.pdf
}
