% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{paris_pra_labour}
\alias{paris_pra_labour}
\title{Paris land value 1990 and 2000}
\usage{
paris_pra_labour(tl0, year = 2000, keep_pras = FALSE)
}
\description{
terre labourable only.
}
\details{
This function produces a csv and tex output table in dropbox at location
\code{"~output/data/tables/paris-pra-prices-2000.csv"}, for example
We take all PRA directly adjacent to paris as well as CEINTURE DE PARIS as
on this map: https://driaaf.ile-de-france.agriculture.gouv.fr/IMG/pdf/RA_R11_cle4ce9b1.pdf
}
