% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{paris_prices_2015}
\alias{paris_prices_2015}
\title{Paris Land Values 2015}
\usage{
paris_prices_2015(keep_pras = FALSE)
}
\description{
values are an AVERAGE over all land types - not labourable vs vignes etc
}
\details{
we collect those from the map at https://driaaf.ile-de-france.agriculture.gouv.fr/IMG/pdf/PRIX_GBP_2015_REGROUP_PRA_cle046232.pdf
}
