% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_density_center}
\alias{plot_density_center}
\title{Plot distance from center}
\usage{
plot_density_center(
  city_name = c("Paris", "Lyon"),
  by_year = TRUE,
  w = 9,
  h = 6
)
}
\arguments{
\item{city_name}{vector of city names}
}
\value{
list with plots
}
\description{
using GHS grid data, plot average density at a certain distance
from the center of a city.
}
\details{
This function uses output of \code{\link{dist_from_center}} and
estimates an exponential decay model via NLS.
}
\references{
https://douglas-watson.github.io/post/2018-09_exponential_curve_fitting/
}
