% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sat_vs_manual}
\alias{plot_sat_vs_manual}
\title{plot manual vs satellite measurements 2015/2016}
\usage{
plot_sat_vs_manual(w = 8.5, h = 5)
}
\description{
makes a plot to compare our manual area measures in 2016 to the satellite measure in 2015
}
