% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_shlomo}
\alias{plot_shlomo}
\title{Compare Area Measures to Shlomo Angel's Results}
\usage{
plot_shlomo(pw = 8, ph = 4.5)
}
\description{
We obtained data from Shlomo Angel's Atlas of Urban Expansion
which contains area estimates fro Paris since 1800. This function
makes a plot combining both measures.
}
