% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top100_cutoff}
\alias{plot_top100_cutoff}
\title{plot top 100 cities densities cutoff sensitivity}
\usage{
plot_top100_cutoff(save = FALSE, w = 9, h = 5)
}
\arguments{
\item{w}{plot width in inches}

\item{h}{plot height in inches}
}
\description{
plot top 100 cities densities cutoff sensitivity
}
