% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top100_densities}
\alias{plot_top100_densities}
\title{plot top 100 cities densities over time
https://github.com/floswald/LandUse.jl/issues/42}
\usage{
plot_top100_densities(
  save = FALSE,
  w = 9,
  h = 5,
  wp = width_paper(),
  hp = height_paper()
)
}
\arguments{
\item{w}{plot width in inches}

\item{h}{plot height in inches

The input data is produced by \code{\link{export_pop_area}}.
We computed average urban density by weighting cities by their 1975 population counts.}
}
\description{
Makes several plots of density and in particular, Figure 4a of the paper.
}
