% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities-defs.R
\name{pop_1950_2}
\alias{pop_1950_2}
\title{Compute French 1954 City Population}
\usage{
pop_1950_2(overwrite = FALSE)
}
\description{
This function aggregates `communes` into greater urban areas according to our own definition.
We establish extent of a city by examining aereal photography and maps from 1950.
Larger cities may contain separate communes as defined by INSEE (and hence, with a separate entry in the census counts).
}
