% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_1950}
\alias{pra_1950}
\title{Match Cities to Prices in 1950}
\usage{
pra_1950(yr = 1950, paris_pras = FALSE)
}
\arguments{
\item{yr}{integer by default 1950. One can choose the relevant year
to consider for the PRA prices. For instance, we tried 1966 once.}

\item{paris_pra}{Boolean by default \code{FALSE} indicating that we do not
want to keep parisian PRAs disaggregated in the data.

outputs to
\verb{data/statistique-agricole/petite-region-agricole/area-pop-prices-1950.csv}}
}
\description{
Similar to \code{\link[=pra_1975]{pra_1975()}}, performs some manual assignment.
adds paris via \code{\link[=paris_pra_1950s_labour]{paris_pra_1950s_labour()}}.
}
