% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_1990}
\alias{pra_1990}
\title{Match Cities to Prices in 1990}
\usage{
pra_1990()
}
\description{
Similar to \code{\link[=pra_2000]{pra_2000()}}, performs some manual assignment.
Adds Paris via \code{\link[=paris_pra_labour]{paris_pra_labour()}}.
}
