% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_2000}
\alias{pra_2000}
\title{Match PRA Prices to Cities in 2000}
\usage{
pra_2000()
}
\description{
Manually associates PRAs to cities, where the mapping is incomplete or ambiguous.
Also takes care of multiple PRAs being associated to a single city via CODGEO,
in which case we take a simple average over all concerned PRA prices.
}
\details{
Input data are from \code{\link[=read_terre_labourable_2000]{read_terre_labourable_2000()}}

Writes output data to
\code{data/statistique-agricole/petite-region-agricole/area-pop-prices-2000-labourable.csv}

Writes the list of cities with multiple PRAs to
\code{data/statistique-agricole/petite-region-agricole/area-pop-prices-2000-labourable-fixlist.csv}
}
