% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{PRA_2015_VVTDiffusion}
\alias{PRA_2015_VVTDiffusion}
\title{petite region agricole 1999-2020 with PRA mapping}
\usage{
PRA_2015_VVTDiffusion(yr = 2015)
}
\description{
petite region agricole 1999-2020 with PRA mapping
}
