% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_IV_neighbors}
\alias{pra_IV_neighbors}
\title{PRA Prices: Neighbor IV Regression}
\usage{
pra_IV_neighbors()
}
\description{
this function performs 2SLS with neighboring PRA prices
as an IV for own PRA price.
}
