% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_manual_check}
\alias{pra_manual_check}
\title{Generate PRA-City mapping for manual check}
\usage{
pra_manual_check()
}
\description{
this function generates a csv file which we manually check for the correct mapping between old and new PRA code.
}
\details{
Our mapping is correct in the vast majority of cases. Some cities have multiple PRAs assigned to them, in particular Paris.
For more info please see \code{\link[=pra_specials]{pra_specials()}} for a precise list.
}
