% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_specials}
\alias{pra_specials}
\title{PRA special cases}
\usage{
pra_specials(overwrite = FALSE)
}
\description{
This returns a list of special cases cities where we make
manual adjustments to the list of surrouding PRAs.
}
\details{
This is for the most part cities where no unique PRA encloses the entire city, which happens for
four cities in total.
}
