% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{pra_stack}
\alias{pra_stack}
\title{Stack City and PRA Data for All Years}
\usage{
pra_stack(topn = 100, overwrite = FALSE)
}
\arguments{
\item{topn}{integer 100 or 200 largest cities to consider}

\item{overwrite}{Boolean TRUE/FALSE whether to overwrite or read from disk}

\item{paris_pras}{Boolean TRUE/FALSE to disaggregate parisian PRAs or not}
}
\description{
Main PRA output function. Builds dataset with different row count
based on how many cities and whether one wants Paris PRAs disaggregated or not.
}
\details{
How does the mapping from PRA prices to cities work? For years up to and
and including 2000, we can rely on a 1:1 mapping of \code{PRA_Code} which maps
city \code{CODGEO} to \code{PRA_Code}. Year 2015 does not have this mapping,
so we first match by department and name and then manually fix the missing data
for each city in function \code{\link[=get_PRA_result_name_1999]{get_PRA_result_name_1999()}}.
\subsection{Function Contents}{
\itemize{
\item Load PRA Data for each year
\item stack all years
\item return from function if we are checking with \code{paris_pra = TRUE}
\item otherwise continue:
\item keep only ranks \code{topn}
\item merge AOC classification by \code{CODGEO}
\item price in Euros
\item merge AU classification by \code{CODGEO}
\item merge Urban wage data by \code{AU2010} (from CASD)
\item compute rank by population of 1975 to get popweights
}
\subsection{Load PRA Data for each year}{

The function has 5 functions that create data (4 in the \code{topn=200} sample).
Additionally the user can choose to output Paris PRAs in a disaggregated way
for checking with argument \code{paris_pra}.
\enumerate{
\item \code{\link[=pra_1950]{pra_1950()}}
\item \code{\link[=pra_1975]{pra_1975()}}
\item \code{\link[=pra_1990]{pra_1990()}}
\item \code{\link[=pra_2000]{pra_2000()}}
\item \code{\link[=add_prices_PRA]{add_prices_PRA()}} with \code{yr = 2015} argument.
}

Please refer to the documentation of each for details.
}

}
}
