% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{PRA_xw}
\alias{PRA_xw}
\title{PRA new-old mapping Crosswalk}
\usage{
PRA_xw()
}
\description{
We recovered this mapping manually, as there is no
official crosswalk from old to new PRA codes available.
}
\details{
The mapping is provided as \verb{PrixCode => PRA_Code}. \code{PrixCode} is column number 7.
}
