% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_petite_1999}
\alias{read_petite_1999}
\title{petite region agricole 1990+}
\usage{
read_petite_1999()
}
\description{
this price data covers 1999 onwards. It has an incomplete mapping to PRA_Code,
which complicates the merge to city data. We first attempt to merge the majority
by department and name, and then fix manually for each city and year the missing
prices (not in this function).
}
\details{
data is from https://agreste.agriculture.gouv.fr/agreste-web/disaron/V_0003/detail/
same data in nicer format here \code{data/statistique-agricole/petite-region-agricole/input/DIFFUSION_VVT_2020_site_DRIAAF_cle0d516d.xlsx}
}
