% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_pra_agreste}
\alias{read_pra_agreste}
\title{petite region agricole 1999-2020}
\usage{
read_pra_agreste(overwrite = FALSE)
}
\description{
Refernce: https://agreste.agriculture.gouv.fr/agreste-web/methodon/Z.1/!searchurl/listeTypeMethodon/
}
\details{
function is deprecated as we use \code{\link[=PRA_xw]{PRA_xw()}} now.

data is from https://agreste.agriculture.gouv.fr/agreste-web/disaron/V_0003/detail/
same data in nicer format here \code{data/statistique-agricole/petite-region-agricole/input/DIFFUSION_VVT_2020_site_DRIAAF_cle0d516d.xlsx}
}
