% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_terre_labourable_1990}
\alias{read_terre_labourable_1990}
\title{terre labourable 1990 - libre a la vente}
\usage{
read_terre_labourable_1990()
}
\description{
read raw data from pdf at \code{statistique-agricole/petite-region-agricole/input/1990-labourables.pdf}
}
