% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_terre_labourable_1999_2000}
\alias{read_terre_labourable_1999_2000}
\title{terre labourable 2000 - libre a la vente}
\usage{
read_terre_labourable_1999_2000()
}
\description{
this is year 2000 Sales prices for \emph{terre labourable}.
read raw data from pdf at \code{statistique-agricole/petite-region-agricole/input/1999-2000-terre-labourables.pdf}
}
