% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_terre_labourable_2000}
\alias{read_terre_labourable_2000}
\title{PRA Arable Land prices for 2000}
\usage{
read_terre_labourable_2000(overwrite = FALSE, paris_pras = FALSE)
}
\description{
Adds Paris via \code{\link[=paris_pra_labour]{paris_pra_labour()}}.
}
