% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{read_terre_labourable_louee_1999_2000}
\alias{read_terre_labourable_louee_1999_2000}
\title{terre labourable 2000 - louee}
\usage{
read_terre_labourable_louee_1999_2000()
}
\description{
from pdf at \code{statistique-agricole/petite-region-agricole/input/1999-2000-terre-labourables-louees.pdf}
}
