% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{relative_pop_area}
\alias{relative_pop_area}
\title{Show Relative population and area}
\usage{
relative_pop_area(cities = top5now(), overwrite = FALSE)
}
\description{
output table for top5 cities with relative population and area to Paris.
}
