% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{table2}
\alias{table2}
\title{Produce Table 2 in Main Text}
\usage{
table2(save = FALSE)
}
\description{
table with 2 columns, model and data, for
regression of log city density on log price of surrounding agricultural
area. The data for the model regression is produced by the function
\code{plot_het} and is part of the corresponding julia package available
\href{https://github.com/floswald/LandUse.jl}{here}.
}
