% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{tableA3}
\alias{tableA3}
\title{Table A.3: Urban Density and Land Values}
\usage{
tableA3(writedisk = FALSE)
}
\description{
This is a more extensive version of \code{\link[=table2]{table2()}}. OLS and IV results with varying controls.
}
\details{
We add 2 columns for both model and data, one without any controls, and one with city FEs.
The function also writes a table to disk for R3 who wanted to see more specifications.
}
