% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petite-region-agricole.R
\name{tableA5_A6}
\alias{tableA5_A6}
\title{Tables A.5 and A.6: Urban Density and Land Values with Shift Share IV}
\usage{
tableA5_A6(writedisk = FALSE)
}
\description{
Same exercise as \code{\link[=tableA3]{tableA3()}} but with shift share IV instead of wheat yields
}
\details{
Produces one tables for first (A5) and second (A6) stages.
}
