% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHS.R
\name{measure_cities}
\alias{measure_cities}
\title{take each cities cropped raster and
measure total pop and built up}
\usage{
measure_cities(overwrite = FALSE, cutoff = 30)
}
\description{
take each cities cropped raster and
measure total pop and built up
}
